+++ src/eval.c
@@ -3139,7 +3139,7 @@ eval_index(
 		    if (n1 >= len || n2 < 0 || n1 > n2)
 			s = NULL;
 		    else
-			s = vim_strnsave(s + n1, (int)(n2 - n1 + 1));
+			s = vim_strnsave(s + n1, n2 - n1 + 1);
 		}
 		else
 		{
